Refactored leave approval screen into 2 widgets
